Benchmark script: PG part can now be disabled. Reran benchmarks, updated readme accordingly.

+++ vignettes/benchmarking.R
@@ -1,7 +1,7 @@
 ####################################################
 ### NOTE
 # For the PostGIS part of this script to run, you require
-# a PostGIS (>= 2.0) enabeld PostgreSQL DB (>=9.5) named
+# an empty, PostGIS (>= 2.0) enabeld PostgreSQL DB (>=9.5) named
 # 'gisdb' on your system, accessible via port 5432.
 ####################################################
 
@@ -16,14 +16,21 @@ library(rbenchmark)
 library(RPostgreSQL)
 library(rgdal)
 
+TEST_PGIS <- FALSE  # Switch off if no PostGIS DB available
+PGIS_USER <- "hunzikp" # Adjust accordingly
+PGIS_PW <- "gotthard"
+
 ####################################################
 ### Setup PGSQL connection
 ####################################################
 
-drv <- dbDriver("PostgreSQL")
-con <- dbConnect(drv, dbname = "gisdb",
-                 host = "localhost", port = 5432,
-                 user = "hunzikp", password = "XXXXX")
+if (TEST_PGIS) {
+  drv <- dbDriver("PostgreSQL")
+  con <- dbConnect(drv, dbname = "gisdb",
+                   host = "localhost", port = 5432,
+                   user = PGIS_USER, password = PGIS_PW)
+}
+
 
 ####################################################
 ### Make testing data
@@ -40,16 +47,6 @@ res <- 1/dim
 large.vx <- velox(large.mat, extent, res, crs="")
 large.rl <- raster(large.mat, xmn=extent[1], xmx=extent[2], ymn=extent[3], ymx=extent[4])
 
-## Import raster into PostGIS and index
-dbSendQuery(con, "DROP TABLE IF EXISTS raster.large;")
-dbSendQuery(con, "DROP SCHEMA IF EXISTS raster;")
-dbSendQuery(con, "CREATE SCHEMA raster;")
-large.vx$write("vignettes/large.tif")
-system("raster2pgsql -s 4326 -I -C -M vignettes/large.tif -F -t 100x100 raster.large > importlarge.sql")
-system("psql -U hunzikp -d gisdb -f importlarge.sql -h localhost -p 5432 -w")
-file.remove("importlarge.sql")
-file.remove("vignettes/large.tif")
-
 ## Create SPDF
 n <- 10
 set.seed(0)
@@ -59,17 +56,27 @@ spol <- gBuffer(sp, width=0.05, byid=TRUE)
 spdf <- SpatialPolygonsDataFrame(spol, data.frame(id=1:length(spol)), FALSE)
 writeOGR(spdf, "vignettes", "spdf", driver="ESRI Shapefile")
 
+## Load data into PostGIS
+if (TEST_PGIS) {
+  ## Import raster into PostGIS and index
+  dbSendQuery(con, "DROP TABLE IF EXISTS raster.large;")
+  dbSendQuery(con, "DROP SCHEMA IF EXISTS raster;")
+  dbSendQuery(con, "CREATE SCHEMA raster;")
+  large.vx$write("vignettes/large.tif")
+  system("raster2pgsql -s 4326 -I -C -M vignettes/large.tif -F -t 100x100 raster.large > importlarge.sql")
+  system(paste0("psql -U ", PGIS_USER," -d gisdb -f importlarge.sql -h localhost -p 5432 -w"))
+  file.remove("importlarge.sql")
+  file.remove("vignettes/large.tif")
 
-## Import SPDF into PostGIS
-dbSendQuery(con, "DROP TABLE IF EXISTS vector.spdf;")
-dbSendQuery(con, "DROP SCHEMA IF EXISTS vector;")
-dbSendQuery(con, "CREATE SCHEMA vector;")
-system("shp2pgsql -s 4326 -I  vignettes/spdf.shp vector.spdf > importspdf.sql")
-system("psql -U hunzikp -d gisdb -f importspdf.sql -h localhost -p 5432 -w")
-file.remove("importspdf.sql")
-system("rm vignettes/spdf.*")
-
-
+  ## Import SPDF into PostGIS
+  dbSendQuery(con, "DROP TABLE IF EXISTS vector.spdf;")
+  dbSendQuery(con, "DROP SCHEMA IF EXISTS vector;")
+  dbSendQuery(con, "CREATE SCHEMA vector;")
+  system("shp2pgsql -s 4326 -I  vignettes/spdf.shp vector.spdf > importspdf.sql")
+  system(paste0("psql -U ", PGIS_USER, " -d gisdb -f importspdf.sql -h localhost -p 5432 -w"))
+  file.remove("importspdf.sql")
+  system("rm vignettes/spdf.*")
+}
 
 
 ####################################################
@@ -78,13 +85,13 @@ system("rm vignettes/spdf.*")
 
 png("vignettes/benchmark.png", width=800, height=600)
 layout(matrix(1:6, 2, 3, TRUE))
-bench <- FALSE  # Turn off if only plotting required
+
 
 ####################################################
 ### Benchmark extraction
 ####################################################
 
-if (bench) {
+if (TEST_PGIS) {
 
   extract.postgis <- function() {
     query <- "SELECT
@@ -105,30 +112,45 @@ if (bench) {
                      postgis=extract.postgis(),
                      replications=10,
                      columns=c("test", "replications", "elapsed", "relative"))
+
+  ex.bm <- ex.bm[order(ex.bm$relative),]
+  par(mar=c(5.1, 4.1, 7.1, 2.1))
+  barplot(ex.bm$relative, names.arg=ex.bm$test,
+          ylab="Relative Time Elapsed", cex.main=1.5)
+  mtext("extract", line=5, cex=1.5, font=2)
+  mtext(paste("velox ", round(ex.bm$relative[ex.bm$test=="raster"], 1), "x faster than raster", sep=""), line=3, cex=1.25)
+  mtext(paste("velox ", round(ex.bm$relative[ex.bm$test=="postgis"], 1), "x faster than PostGIS", sep=""), line=1, cex=1.25)
+
+} else {
+
+  ex.bm <- benchmark(velox=large.vx$extract(spdf, sum),
+                     raster=extract(large.rl, spdf, fun=sum),
+                     replications=10,
+                     columns=c("test", "replications", "elapsed", "relative"))
+
+  ex.bm <- ex.bm[order(ex.bm$relative),]
+  par(mar=c(5.1, 4.1, 7.1, 2.1))
+  barplot(ex.bm$relative, names.arg=ex.bm$test,
+          ylab="Relative Time Elapsed", cex.main=1.5)
+  mtext("extract", line=5, cex=1.5, font=2)
+  mtext(paste("velox ", round(ex.bm$relative[ex.bm$test=="raster"], 1), "x faster than raster", sep=""), line=3, cex=1.25)
 }
 
-ex.bm <- ex.bm[order(ex.bm$relative),]
-par(mar=c(5.1, 4.1, 7.1, 2.1))
-barplot(ex.bm$relative, names.arg=ex.bm$test,
-        ylab="Relative Time Elapsed", cex.main=1.5)
-mtext("extract", line=5, cex=1.5, font=2)
-mtext(paste("velox ", round(ex.bm$relative[ex.bm$test=="raster"], 1), "x faster than raster", sep=""), line=3, cex=1.25)
-mtext(paste("velox ", round(ex.bm$relative[ex.bm$test=="postgis"], 1), "x faster than PostGIS", sep=""), line=1, cex=1.25)
 
 ####################################################
 ### Benchmark aggregation
 ####################################################
 
-if (bench) {
-  agg.velox <- function() {
-    agg.vx <- large.vx$copy()
-    agg.vx$aggregate(c(16,16), "mean")
-  }
-  ag.bm <- benchmark(velox=agg.velox(),
-                     raster=aggregate(large.rl, fact=c(16,16), fun=mean, expand=FALSE),
-                     replications=10,
-                     columns=c("test", "replications", "elapsed", "relative"))
+
+agg.velox <- function() {
+  agg.vx <- large.vx$copy()
+  agg.vx$aggregate(c(16,16), "mean")
 }
+ag.bm <- benchmark(velox=agg.velox(),
+                   raster=aggregate(large.rl, fact=c(16,16), fun=mean, expand=FALSE),
+                   replications=10,
+                   columns=c("test", "replications", "elapsed", "relative"))
+
 ag.bm <- ag.bm[order(ag.bm$relative),]
 par(mar=c(5.1, 4.1, 6.1, 2.1))
 barplot(ag.bm$relative, names.arg=ag.bm$test,
@@ -141,17 +163,17 @@ mtext(paste("velox ", round(ag.bm$relative[ag.bm$test=="raster"], 1), "x faster 
 ### Benchmark cropping
 ####################################################
 
-if (bench) {
-  crop.sp <- spdf[1,]
-  crop.velox <- function() {
-    crop.vx <- large.vx$copy()
-    crop.vx$crop(crop.sp)
-  }
-  cr.bm <- benchmark(velox=crop.velox(),
-                     raster=crop(large.rl, crop.sp),
-                     replications=10,
-                     columns=c("test", "replications", "elapsed", "relative"))
+
+crop.sp <- spdf[1,]
+crop.velox <- function() {
+  crop.vx <- large.vx$copy()
+  crop.vx$crop(crop.sp)
 }
+cr.bm <- benchmark(velox=crop.velox(),
+                   raster=crop(large.rl, crop.sp),
+                   replications=10,
+                   columns=c("test", "replications", "elapsed", "relative"))
+
 cr.bm <- cr.bm[order(cr.bm$relative),]
 par(mar=c(5.1, 4.1, 6.1, 2.1))
 barplot(cr.bm$relative, names.arg=cr.bm$test,
@@ -164,18 +186,17 @@ mtext(paste("velox ", round(cr.bm$relative[cr.bm$test=="raster"], 1), "x faster 
 ### Benchmark median focal
 ####################################################
 
-if (bench) {
-  mfocal.velox <- function() {
-    mfocal.vx <- large.vx$copy()
-    mfocal.vx$medianFocal(5, 5)
-  }
-  weights <- matrix(1, 5, 5)
-
-  mf.bm <- benchmark(velox=mfocal.velox(),
-                     raster=focal(large.rl, w=weights, fun=median, na.rm=TRUE, pad=TRUE, padValue=NA),
-                     replications=10,
-                     columns=c("test", "replications", "elapsed", "relative"))
+mfocal.velox <- function() {
+  mfocal.vx <- large.vx$copy()
+  mfocal.vx$medianFocal(5, 5)
 }
+weights <- matrix(1, 5, 5)
+
+mf.bm <- benchmark(velox=mfocal.velox(),
+                   raster=focal(large.rl, w=weights, fun=median, na.rm=TRUE, pad=TRUE, padValue=NA),
+                   replications=10,
+                   columns=c("test", "replications", "elapsed", "relative"))
+
 mf.bm <- mf.bm[order(mf.bm$relative),]
 par(mar=c(5.1, 4.1, 6.1, 2.1))
 barplot(mf.bm$relative, names.arg=mf.bm$test,
@@ -190,18 +211,18 @@ mtext(paste("velox ", round(mf.bm$relative[mf.bm$test=="raster"], 1), "x faster 
 ### Benchmark sum focal
 ####################################################
 
-if (bench) {
-  weights <- matrix(1, 5, 5)
-  sfocal.velox <- function() {
-    sfocal.vx <- large.vx$copy()
-    sfocal.vx$sumFocal(weights, 1)
-  }
 
-  sf.bm <- benchmark(velox=sfocal.velox(),
-                     raster=focal(large.rl, w=weights, fun=mean, na.rm=TRUE, pad=TRUE, padValue=NA),
-                     replications=10,
-                     columns=c("test", "replications", "elapsed", "relative"))
+weights <- matrix(1, 5, 5)
+sfocal.velox <- function() {
+  sfocal.vx <- large.vx$copy()
+  sfocal.vx$sumFocal(weights, 1)
 }
+
+sf.bm <- benchmark(velox=sfocal.velox(),
+                   raster=focal(large.rl, w=weights, fun=mean, na.rm=TRUE, pad=TRUE, padValue=NA),
+                   replications=10,
+                   columns=c("test", "replications", "elapsed", "relative"))
+
 sf.bm <- sf.bm[order(sf.bm$relative),]
 par(mar=c(5.1, 4.1, 6.1, 2.1))
 barplot(sf.bm$relative, names.arg=sf.bm$test,
@@ -215,13 +236,13 @@ mtext(paste("velox ", round(sf.bm$relative[sf.bm$test=="raster"], 1), "x faster 
 ### Benchmark rasterize
 ####################################################
 
-if (bench) {
-  rs.bm <- benchmark(velox=large.vx$rasterize(spdf, "id", 1, background=0),
-                     raster=rasterize(spdf, large.rl, "id", background=0),
-                     replications=10,
-                     columns=c("test", "replications", "elapsed", "relative"))
-  rs.bm
-}
+
+rs.bm <- benchmark(velox=large.vx$rasterize(spdf, "id", 1, background=0),
+                   raster=rasterize(spdf, large.rl, "id", background=0),
+                   replications=10,
+                   columns=c("test", "replications", "elapsed", "relative"))
+rs.bm
+
 rs.bm <- rs.bm[order(rs.bm$relative),]
 par(mar=c(5.1, 4.1, 6.1, 2.1))
 barplot(rs.bm$relative, names.arg=rs.bm$test,
